feat: update expression editor to use surveyExpressionCategories and enhance argument handling in various components
