+++ components/survey-editor/components/ItemEditor/editor/components/content-editor/response-editors/exp-arg-editor-for-num.tsx
@@ -1,5 +1,5 @@
 import ExpArgEditor from '@/components/expression-editor/exp-arg-editor';
-import { supportedBuiltInSlotTypes, surveyEngineCategories, surveyEngineRegistry } from '@/components/expression-editor/registries/surveyEngineRegistry';
+import { supportedBuiltInSlotTypes, surveyExpressionCategories, surveyEngineRegistry } from '@/components/expression-editor/registries/surveyEngineRegistry';
 import { ExpressionArg } from '@/components/expression-editor/utils';
 import React, { useEffect } from 'react';
 
@@ -34,7 +34,7 @@ const ExpArgEditorForNum: React.FC<ExpArgEditorForNumProps> = (props) => {
             expRegistry={{
                 expressionDefs: surveyEngineRegistry,
                 builtInSlotTypes: supportedBuiltInSlotTypes,
-                categories: surveyEngineCategories,
+                categories: surveyExpressionCategories,
             }}
             currentIndex={0}
             slotDef={{
